+++ src/app/admin/dashboard/dashboard.component.ts
@@ -90,10 +90,11 @@ export class DashboardComponent extends HandleSubscription implements OnInit {
       ],
     },
     {
-      title: 'Account settings',
+      title: 'Wallet settings',
       description: '',
       link: '/admin/dashboard/account',
       values: [
+        {name: 'Email & password', icon: 'assets/images/preferences.svg'},
         {name: 'Email & password', icon: 'assets/images/preferences.svg'},
         {name: 'Referral links', icon: 'assets/images/preferences.svg'},
       ],