+++ src/app/settings/general-settings/account-settings/account-settings.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { LocalStorageUser } from 'models/user.model'
 
 @Component({
   selector: 'app-account-settings',