+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.11.0] - 2021-12-31
+### Added
+- Connecting your account with the cryptocurrency wallet
+- Log in to your account with the cryptocurrency wallet
+- Auto registration
+- Auto withdrawal
+### Fixed
+- Remove register link in private mode
+
 ## [1.10.8] - 2021-12-10
 ### Fixed
 - CSS & impersonated mode
@@ -331,7 +340,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Publisher features (Sites & AdUnits)
 
 
-[Unreleased]: https://github.com/adshares/adpanel/compare/v1.10.7...develop
+[Unreleased]: https://github.com/adshares/adpanel/compare/v1.11.0...develop
+[1.11.0]: https://github.com/adshares/adpanel/compare/v1.10.8...v1.11.0
+[1.10.8]: https://github.com/adshares/adpanel/compare/v1.10.7...v1.10.8
 [1.10.7]: https://github.com/adshares/adpanel/compare/v1.10.6...v1.10.7
 [1.10.6]: https://github.com/adshares/adpanel/compare/v1.10.5...v1.10.6
 [1.10.5]: https://github.com/adshares/adpanel/compare/v1.10.4...v1.10.5